Modify yaml code for dedicated ingress-controller
This change demonstrates a "hello world" deployment that uses a
dedicated ingress controller. This is going to be the default from now
on, so the user guide tutorial will also need to be modified so that a
dedicated ingress controller is created with the tutorial namespace.

This change should not be deployed until we have modified the cloud
platform CLI so that an ingress controller is created by default, with
every new namespace.

related to
https://github.com/ministryofjustice/cloud-platform/issues/2224